Move vote delta functions to new module
